LPS-117661 Deleting base class and moving remaining logic to the AudioConverter
